Corregido

+++ PerezSantiagoS4C4/factorial.cpp
@@ -6,7 +6,7 @@ int main(){
 int n;
 cout<<"Introduzca el valor de n";
 cin>>n;
-int result=1;
+double result=1;
 for (int i=1;i<=n;i++){
     result= i*result;}
 cout<<result<< endl;
